Minor: learned I cannot technically fork my own repo
I wanted to fork
so that multiple ‘releases’ (each associated with one paper)
could be archived as separate Zenodo umbrella records.
I do not want releases to be versions of a single record,
because that would make things complicated
if I do in fact make multiple records for the files
associated with a single paper.