Update fasta file names, `cat` references together
Remove related fossil commands,
used in the past to export CabLCV and phiX174 sequences.
Other virus sequences were not used for the paper
but are useful for assessing contamination.

Current available TAIR10 reference fasta
has an updated mitochondrial sequence
and is compressed
(in contrast to the previous version,
 which was used for the paper).

GenBank CP001637 fasta file checksum changed
because there is only one newline at the end of the file
when saved with Entrez e-utilities.